2020 AoC Day 23
Part 1 was fine, but was initially waiting for some really poorly
performing versions of Part 2 to finish (like using a deque). Finally
the light went on and I realized I didn't actually *need* index based
access, just access to the next one clockwise based on the label. With
that selected, the rest of the algorithm falls out in a pretty
straight-forward way.

This was also fun to optimize a bit, since we're running 10^7 iterations
of pure Python code, some small optimizations in choice of operation (or
using array vs. dict) can have a big impact.