2018 AOC Day 22
Finally finished this one up. Took way to long to debug part 2, even
though I had the right algorithm from the start. Just need to actually
fully read the directions--this time it was the fact that when changing
equipment, the change needs to be valid for both current and next
position. Also needed to bump the cache size to avoid blowing the
recursion limit.