Simplify 2021 Day 8 even further
It turns out you only need two segments (b, e) as well as the pair of
(a, c) or (b, d), either of which can be found from their counts. This
allows you to do the necessary separation among (2, 3, 5) and (0, 6, 9).
This eliminate the need to identify the four up front, as well as a lot
of unnecessary book-keeping.