Even more simplification of 2021 Day 8
We can refactor to do the decoding in one fell swoop, no need to
generate a full mapping and apply it. Instead, just use the list of full
patterns to find 1 and 4, and hence cf/bd. After that, we only need to
decode the actual digits given, not all 10 possible patterns.

This eliminates generating the data structure for the mapping, sorting
of characters in all the patterns, and processes fewer of the patterns.
Not a noticable time change due to the size of the dataset, but
definitely much less work.