2021 AOC Day 13
This was pretty easy if you grok what the folds do to the math of the
coordinates. Would have been a bit quicker on part 1 if I had read it
properly (yet again) and seen the "after 1 fold". Happy, though, that I
was ready to go with the "decode the text" portion.