Add alternative solution runner
This simply imports the solution and tries to execute the `run()`
function with the data for the day. This should simplify running with
different data for different accounts. It's also much simpler code and
promotes running the test data with the same processing as the main
solution--just swap them out.

Also with this set up, running the solutions with the aoc runner doesn't
go through running with the sample data.

This will also help deal with the fact that the aoc runner swallows
issues where the solution now produces a different answer than is
cached.