AOC 2023 Day 18
Initally solved Part 1 walking the path and doing a flood fill, but this
blows up with part 2. Ended up applying area formula and then
indpedendently finding Pick's theorem, which in the end makes this a
simple solution. A little proud of putting the part1/part2 split inside
the iterator rather than any of the solution code.